% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/merExtract.R
\name{Extracts random effects}
\alias{Extracts random effects}
\alias{REextract}
\title{REextract}
\usage{
REextract(mod)
}
\arguments{
\item{mod}{a merMod object from the lme4 package}
}
\value{
a data frame with random effects from lme4
}
\description{
Extracts random effect terms from an lme4 model
}

