% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{famlink}
\alias{famlink}
\title{Find link function family}
\usage{
famlink(object, resp = object@resp)
}
\arguments{
\item{object}{a merMod object}

\item{resp}{the response vector}
}
\value{
the link function and family
}
\description{
Find link function family
}
