% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merExtract.R
\name{FEsim}
\alias{FEsim}
\title{Simulate fixed effects from merMod}
\usage{
FEsim(mod, nsims = 100)
}
\arguments{
\item{mod}{a merMod object from the lme4 package}

\item{nsims}{number of simulations to use}
}
\value{
a data frame with distribution of fixed effect parameters
}
\description{
Simulate fixed effects from merMod object posterior distributions
}
\details{
Use the Gelman sim technique to build fixed effect estimates and
confidence intervals. Uses the sim function in the arm package
}

