% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{fetch.merMod.msgs}
\alias{fetch.merMod.msgs}
\title{Extract all warning msgs from a merMod object}
\usage{
fetch.merMod.msgs(x)
}
\arguments{
\item{x}{a merMod object}
}
\description{
Extract all warning msgs from a merMod object
}
