% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{formulaBuild}
\alias{formulaBuild}
\title{Clean formula}
\usage{
formulaBuild(model)
}
\arguments{
\item{model}{a merMod object from lme4}
}
\value{
a formula object
}
\description{
a function to modify the formula for a merMod object to create
a model matrix with all predictor terms in both the group level and fixed
effect level
}
\keyword{internal}
