% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merData.R
\name{hasWeights}
\alias{hasWeights}
\title{Identify if a merMod has weights}
\usage{
hasWeights(merMod)
}
\arguments{
\item{merMod}{the merMod object to test for weights}
}
\value{
TRUE or FALSE for whether the model has weights
}
\description{
Identify if a merMod has weights
}
