% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ICC}
\alias{ICC}
\title{Calculate the intraclass correlation using mixed effect models}
\usage{
ICC(outcome, group, data, subset = NULL)
}
\arguments{
\item{outcome}{a character representing the variable of the outcome}

\item{group}{a character representing the name of the grouping term}

\item{data}{a data.frame}

\item{subset}{an optional subset}
}
\value{
a numeric for the intraclass correlation
}
\description{
Calculate the intraclass correlation using mixed effect models
}
\examples{
data(sleepstudy)
ICC(outcome = "Reaction", group = "Subject", data = sleepstudy)
}
