% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merList.R
\name{lmerModList}
\alias{bglmerModList}
\alias{blmerModList}
\alias{glmerModList}
\alias{lmerModList}
\title{Apply a multilevel model to a list of data frames}
\usage{
lmerModList(formula, data, parallel = NULL, ...)

blmerModList(formula, data, parallel = NULL, ...)

glmerModList(formula, data, parallel = NULL, ...)

bglmerModList(formula, data, parallel = NULL, ...)
}
\arguments{
\item{formula}{a formula to pass through compatible with merMod}

\item{data}{a list object with each element being a data.frame}

\item{parallel}{logical, should the models be run in parallel?}

\item{...}{additional arguments to pass to the estimating function}
}
\value{
a list of fitted merMod objects of class merModList

a merModList

a merModList

a merModList
}
\description{
Apply a multilevel model to a list of data frames

Apply a Bayesian multilevel model to a list of data frames

Apply a generalized linear multilevel model to a list of data frames

Apply a Bayesian generalized linear multilevel model to a list of data frames
}

