% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merTools-package.r
\docType{package}
\name{merTools-package}
\alias{merTools}
\alias{merTools-package}
\title{merTools: Provides methods for extracting and exploring results from merMod
objects in the lme4 package.}
\description{
The merTools package contains convenience tools for extracting useful
information from and exploring the implications of merMod objects created by
the lme4 package.  These convenience functions are especially useful for
merMod objects that take a long time to estimate due to their complexity or
because they are estimated on very large samples.
}
\details{
See the vignettes for usage examples
}
\section{merMod extraction/utility functions}{


\itemize{
  \item \code{\link{fastdisp}}
  \item \code{\link{superFactor}}
  \item \code{\link{REextract}}
  \item \code{\link{REsim}}
  \item \code{\link{FEsim}}
  \item \code{\link{RMSE.merMod}}
  \item \code{\link{thetaExtract}}
  \item \code{\link{REquantile}}
}
}

\section{merMod exploration functions}{


\itemize{
  \item \code{\link{plotREsim}}
  \item \code{\link{plotFEsim}}
  \item \code{\link{plotREimpact}}
  \item \code{\link{draw}}
  \item \code{\link{wiggle}}
  \item \code{\link{subBoot}}
  \item \code{\link{predictInterval}}
  \item \code{\link{expectedRank}}
  \item \code{\link{REimpact}}
  \item \code{\link{shinyMer}}
}
}

\section{Influences and acknowledgements}{


The simulation approach used by \code{\link{predictInterval}} to propagate
uncertainty from both the fixed and random effects draws on the framework
described by Gelman and Hill (2007) and implemented in the \pkg{arm}
package's \code{sim()} function. The package is built on top of the model
objects fit by \pkg{lme4} (Bates et al. 2015) and is intended to complement
\code{lme4::bootMer()} by offering a faster, simulation-based alternative
for large models. We are grateful to the authors and maintainers of those
packages.
}

\references{
Gelman, A. and Hill, J. (2007). \emph{Data Analysis Using Regression and
Multilevel/Hierarchical Models}. Cambridge University Press.

Bates, D., Maechler, M., Bolker, B., and Walker, S. (2015). Fitting Linear
Mixed-Effects Models Using lme4. \emph{Journal of Statistical Software},
67(1), 1-48. \doi{10.18637/jss.v067.i01}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jknowles.github.io/merTools/}
  \item \url{https://github.com/jknowles/merTools}
  \item Report bugs at \url{https://github.com/jknowles/merTools/issues}
}

}
\author{
\strong{Maintainer}: Jared E. Knowles \email{jared@civilytics.com}

Authors:
\itemize{
  \item Jared E. Knowles \email{jared@civilytics.com}
  \item Carl Frederick \email{carlbfrederick@gmail.com}
}

Other contributors:
\itemize{
  \item Alex Whitworth \email{whitworth.alex@gmail.com} [contributor]
}

}
\keyword{internal}
