% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mkNewReTrms}
\alias{mkNewReTrms}
\title{Utility function to make RE terms objects}
\usage{
mkNewReTrms(object, newdata, re.form = NULL, na.action = na.pass,
  allow.new.levels = FALSE)
}
\description{
Utility function to make RE terms objects
}
\keyword{internal}

