% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merList.R
\name{modelFixedEff}
\alias{modelFixedEff}
\title{Extract averaged fixed effect parameters across a list of merMod objects}
\usage{
modelFixedEff(modList)
}
\arguments{
\item{an}{object of class merModList}
}
\value{
a data.frame of the averaged fixed effect parameters
}
\description{
Extract averaged fixed effect parameters across a list of merMod objects
}

