% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvrnormArma}
\alias{mvrnormArma}
\title{Faster multivariate normal sampling}
\usage{
mvrnormArma(n, mu, sigma)
}
\arguments{
\item{n}{number of samples to draw}

\item{mu}{vector of means to use for the distribution}

\item{sigma}{matrix of sigma values for distribution}
}
\description{
Speed up sampling from multivariate normal using C++
}

