% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merPlots.R
\name{plotFEsim}
\alias{plotFEsim}
\title{Plot the results of a simulation of the fixed effects}
\usage{
plotFEsim(data, scale = 1.96, var = "median_eff", sd = "sd_eff",
  intercept = FALSE, sigmaScale = NULL, oddsRatio = FALSE)
}
\arguments{
\item{data}{a data.frame generated by \code{\link{FEsim}} with simulations of
the fixed effects of a \code{\link{merMod}}}

\item{scale}{a factor to multiply the effect standard deviation by, default is 1.96 for
95\% confidence intervals}

\item{var}{a character value indicating the variable name in data of the
midpoint of the estimated interval, e.g. "mean" or "median"}

\item{sd}{a character value indicating the variable name in data of the
variance of the estimated interval, e.g."sd"}

\item{intercept}{logical, should the intercept be included, default is FALSE}

\item{sigmaScale}{a numeric value to divide the estimate and the standard
deviation by in the case of doing an effect size calculation}

\item{oddsRatio}{logical, should the parameters be converted to odds ratios
before plotting}
}
\value{
a ggplot2 plot of the coefficient effects
}
\description{
Plot the simulated fixed effects on a ggplot2 chart
}

