% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merPlots.R
\name{plotREsim}
\alias{plotREsim}
\title{Plot the results of a simulation of the random effects}
\usage{
plotREsim(data, scale = 1.96, var = "median_eff", sd = "sd_eff",
  sigmaScale = NULL, oddsRatio = FALSE, labs = FALSE)
}
\arguments{
\item{data}{a data.frame generated by \code{\link{REsim}} with simulations of
the random effects of a \code{\link{merMod}}}

\item{scale}{a factor to multiply the effect standard deviation by, default is 1.96 for
95\% confidence intervals}

\item{var}{a character value indicating the variable name in data of the
midpoint of the estimated interval, e.g. "mean" or "median"}

\item{sd}{a character value indicating the variable name in data of the
variance of the estimated interval, e.g. "sd"}

\item{sigmaScale}{a numeric value to divide the estimate and the standard
deviation by in the case of doing an effect size calculation}

\item{oddsRatio}{logical, should the parameters be converted to odds ratios
before plotting}

\item{labs}{logical, include the labels of the groups on the x-axis}
}
\value{
a ggplot2 plot of the coefficient effects
}
\description{
Plot the simulated random effects on a ggplot2 chart
}

