% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merList.R
\name{print.merModList}
\alias{print.merModList}
\title{Print the results of a merMod list}
\usage{
\method{print}{merModList}(x, ...)
}
\arguments{
\item{x}{a modList of class merModList}

\item{...}{additional arguments}
}
\value{
summary content printed to console
}
\description{
Print the results of a merMod list
}

