% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merList.R
\name{print.summary.merModList}
\alias{print.summary.merModList}
\title{Print the summary of a merMod list}
\usage{
\method{print}{summary.merModList}(x, ...)
}
\arguments{
\item{x}{a summary of amerModList object}

\item{...}{additional arguments}
}
\value{
summary content printed to console
}
\description{
Print the summary of a merMod list
}
