% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm.R
\name{rccp_rmvnorm}
\alias{rccp_rmvnorm}
\alias{rcpp_rmvnorm}
\title{Fast rcpp rmvnorm sampler}
\usage{
rcpp_rmvnorm(n, mu, sigma)
}
\arguments{
\item{n}{number of samples}

\item{mu}{vector of means}

\item{sigma}{sigma matrix}
}

