% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcppeigen_get_chol}
\alias{rcppeigen_get_chol}
\title{find the Cholesky decomposition of a matrix with RcppEigen}
\usage{
rcppeigen_get_chol(A)
}
\arguments{
\item{A}{a matrix to decompose}
}
\description{
Speed up Cholesky SPD matrix decomposition
}

