% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merList.R
\name{REcorrExtract}
\alias{REcorrExtract}
\title{Extract the correlations between the slopes and the intercepts from a model}
\usage{
REcorrExtract(model)
}
\arguments{
\item{model}{an object that inherits from class merMod}
}
\value{
a numeric vector of the correlations among the effects
}
\description{
Extract the correlations between the slopes and the intercepts from a model
}
\examples{
fm1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
REcorrExtract(fm1)
}
