% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merExtract.R
\name{REextract}
\alias{REextract}
\title{Extracts random effects}
\usage{
REextract(mod)
}
\arguments{
\item{mod}{a merMod object from the lme4 package}
}
\value{
a data frame with random effects from lme4 by group id
}
\description{
Extracts random effect terms from an lme4 model
}

