% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{reTermCount}
\alias{reTermCount}
\title{Count the number of random effect terms}
\source{
From lme4 package
}
\usage{
reTermCount(model)
}
\description{
Count the number of random effect terms
}
\keyword{internal}
