% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{reTermNames}
\alias{reTermNames}
\title{Get names of random effect terms in a model object}
\usage{
reTermNames(model)
}
\arguments{
\item{model}{a merMod object with random effect terms}
}
\value{
a data.frame with rows for each term with columns naming the grouping
term and the effect type
}
\description{
Get names of random effect terms in a model object
}
\keyword{internal}
