% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merData.R
\name{sanitizeNames}
\alias{sanitizeNames}
\title{Clean up variable names in data frames}
\usage{
sanitizeNames(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
a data frame with variable names cleaned to remove factor() construction
}
\description{
Strips out transformations from variable names in data frames
}

