% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyMer.R
\name{shinyMer}
\alias{shinyMer}
\title{Launch a shiny app to explore your merMod interactively}
\usage{
shinyMer(merMod, newdata = NULL)
}
\arguments{
\item{merMod}{An object of class "merMod".}

\item{newdata}{A data.frame to make predictions from (optional). If
  NULL, then the user can only make predictions using the data in
  the frame slot of the merMod object.}
}
\value{
A shiny app
}
\description{
\code{shinyMer} launches a shiny app that allows you to interactively
explore an estimated merMod using functions from \code{merTools}.
}

