% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merData.R
\name{shuffle}
\alias{shuffle}
\title{Randomly reorder a dataframe}
\usage{
shuffle(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame of the same dimensions with the rows reordered
randomly
}
\description{
Randomly reorder a dataframe by row
}
