% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subBoot.R
\name{subBoot}
\alias{subBoot}
\title{Bootstrap a subset of an lme4 model}
\usage{
subBoot(model, n, FUN, B)
}
\arguments{
\item{model}{a valid merMod object}

\item{n}{the number of rows to sample from the original data
in the merMod object}

\item{FUN}{the function to apply to each bootstrapped model}

\item{B}{the number of bootstrap replicates}
}
\value{
a matrix of parameters extracted from each of the B replications
}
\description{
Bootstrap a subset of an lme4 model
}
\details{
This function allows users to estimate parameters of a
large merMod object using bootstraps on a subset of the data.
}

