% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merData.R
\name{subsetList}
\alias{subsetList}
\title{Subset a data.frame using a list of conditions}
\usage{
subsetList(data, list)
}
\arguments{
\item{data}{a data.frame}

\item{list}{a named list of splitting conditions}
}
\value{
a data frame with values that match the conditions in the list
}
\description{
Split a data.frame by elements in a list
}

