% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{sum.mm}
\alias{sum.mm}
\title{Title}
\usage{
\method{sum}{mm}(object, correlation = (p <=
  getOption("lme4.summary.cor.max")), use.hessian = NULL, ...)
}
\arguments{
\item{object}{a merMod object}

\item{correlation}{optional p value}

\item{use.hessian}{logical}

\item{...}{}
}
\value{
a summary of the object
}
\description{
Title
}
