% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merList.R
\name{summary.merModList}
\alias{summary.merModList}
\title{Summarize a merMod list}
\usage{
\method{summary}{merModList}(object, ...)
}
\arguments{
\item{object}{a modList of class merModList}

\item{...}{additional arguments}
}
\value{
a summary object of model information
}
\description{
Summarize a merMod list
}
