% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subBoot.R
\name{thetaExtract}
\alias{thetaExtract}
\title{Extract theta parameters from a merMod model}
\usage{
thetaExtract(model)
}
\arguments{
\item{model}{a valide merMod object}
}
\value{
a vector of the covrariance, theta, parameters from a \code{\link{merMod}}
}
\description{
A convenience function that returns the theta parameters for a
\code{\link{merMod}} obejct.
}
\seealso{
merMod
}

