% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merFastDisplay.R
\name{fastdisp}
\alias{fastdisp}
\title{Display model fit summary of merMod objects, fast}
\usage{
fastdisp(object, ...)
}
\arguments{
\item{object}{a merMod object from the lme4 package}

\item{...}{additional arguments to pass to display function from arm
including number of digits}
}
\value{
A printed summary of a merMod object
}
\description{
A faster version of the display function from arm which does
not refit the model to extract the deviance
}

