% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mkNewReTrms}
\alias{mkNewReTrms}
\title{Utility function to make RE terms objects}
\usage{
mkNewReTrms(object, newdata, re.form = NULL, na.action = na.pass,
  allow.new.levels = FALSE)
}
\arguments{
\item{object}{a model object}

\item{newdata}{a data.frame to build RE terms for}

\item{re.form}{a random effect formula to simulate, generated by
\code{\link{reOnly}}}

\item{na.action}{an object describing how NA values should be handled in newdata}

\item{allow.new.levels}{logical, should new levels be allowed in factor variables}
}
\value{
a random effect terms object for a merMod
}
\description{
Utility function to make RE terms objects
}
\keyword{internal}
