% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merList.R
\name{modelFixedEff}
\alias{modelFixedEff}
\title{Extract averaged fixed effect parameters across a list of merMod objects}
\usage{
modelFixedEff(modList, ...)
}
\arguments{
\item{modList}{an object of class merModList}

\item{...}{additional arguments to pass to \code{\link{tidy}}}
}
\value{
a data.frame of the averaged fixed effect parameters
}
\description{
Extract averaged fixed effect parameters across a list of merMod objects
}
\examples{
sim_list <- replicate(n = 10,
        expr = sleepstudy[sample(row.names(sleepstudy), 180),],
        simplify=FALSE)
fml <- "Reaction ~ Days + (Days | Subject)"
mod <- lmerModList(fml, data = sim_list)
modelFixedEff(mod)
}
