% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merList.R
\name{modelInfo}
\alias{modelInfo}
\title{Extract model information from a merMod}
\usage{
modelInfo(object)
}
\arguments{
\item{object}{a merMod object}
}
\value{
Simple summary information about the object, coefficients, number
of observations, and adjusted R squared
}
\description{
Extract model information from a merMod
}
