% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merList.R
\name{modelRandEffStats}
\alias{modelRandEffStats}
\title{Extract data.frame of random effect statistics from merMod List}
\usage{
modelRandEffStats(modList)
}
\arguments{
\item{modList}{a list of multilevel models}
}
\value{
a data.frame
}
\description{
Extract data.frame of random effect statistics from merMod List
}
