% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merList.R
\name{modelRandEffStats}
\alias{modelRandEffStats}
\title{Extract data.frame of random effect statistics from merMod List}
\usage{
modelRandEffStats(modList)
}
\arguments{
\item{modList}{}
}
\value{
a data.frame
}
\description{
Extract data.frame of random effect statistics from merMod List
}

