% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merPredict.R
\name{predictInterval}
\alias{predictInterval}
\title{Predict from merMod objects with a prediction interval}
\usage{
predictInterval(model, newdata, level = 0.95, nsim = 1000,
  stat = c("median", "mean"), predict.type = c("linear.prediction",
  "probability"), include.resid.var = TRUE)
}
\arguments{
\item{model}{a merMod object from lme4}

\item{newdata}{a data.frame of new data to predict}

\item{level}{the width of the prediction interval}

\item{nsim}{number of simulation samples to construct}

\item{stat}{take the median or mean of simulated intervals}

\item{predict.type}{type of prediction to develop}

\item{include.resid.var}{logical, include or exclude the residual varaince for
linear models}
}
\value{
'newdata' with three columns appended, stat and the lower
        and upper prediction interval boundaries
}
\description{
Predict from merMod objects with a prediction interval
}
\details{
Sampling strategy description.
}

