% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merList.R
\name{print.merModList}
\alias{print.merModList}
\title{Print the results of a merMod list}
\usage{
\method{print}{merModList}(modList, digits = 3)
}
\arguments{
\item{modList}{an object of class merModList}

\item{digits}{a numeric, number of digits to print}
}
\value{
summary content printed to console
}
\description{
Print the results of a merMod list
}

