% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merData.R
\name{randomObs}
\alias{randomObs}
\title{Select a random observation from model data}
\usage{
randomObs(merMod)
}
\arguments{
\item{merMod}{an object of class merMod}
}
\value{
a data frame with a single row for a random observation, but with full
factor levels. See details for more.
}
\description{
Select a random observation from the model frame of a merMod
}
\details{
Each factor variable in the data frame has all factor levels from the
full model.frame stored so that the new data is compatible with predict.merMod
}

