% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merList.R
\name{REsdExtract}
\alias{REsdExtract}
\title{Extract the standard deviation of the random effects from a merMod object}
\usage{
REsdExtract(model)
}
\arguments{
\item{model}{an object that inherits from class merMod}
}
\value{
a numeric vector for standard deviations of the random effects
}
\description{
Extract the standard deviation of the random effects from a merMod object
}

