% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merExtract.R
\name{REsim}
\alias{REsim}
\title{Simulate random effects from merMod}
\usage{
REsim(mod, n.sims = 200, OR = FALSE)
}
\arguments{
\item{mod}{a merMod object from the lme4 package}

\item{n.sims}{number of simulations to use}

\item{OR}{logical, should parameters be converted to odds ratios?}
}
\value{
a data frame with distribution of random effect parameters
}
\description{
Simulate random effects from merMod object posterior distributions
}
\details{
Use the Gelman sim technique to build empirical Bayes estimates.
 Uses the sim function in the arm package
}

