% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merList.R
\name{VarCorr.merModList}
\alias{VarCorr.merModList}
\title{Extract the variances and correlations for random effects from a merMod list}
\usage{
\method{VarCorr}{merModList}(x, sigma = 1, rdig = 3)
}
\arguments{
\item{x}{for \code{VarCorr}: a fitted model object, usually an object inheriting from
    class \code{\linkS4class{merMod}}. For \code{as.data.frame}, a
    \code{VarCorr.merMod} object returned from \code{VarCorr}.}

\item{sigma}{an optional numeric value used as a multiplier for the
    standard deviations.}
}
\value{
a list with two elements "stddev" and "correlation" for the standard
deviations and correlations averaged across models in the list
}
\description{
Extract the variances and correlations for random effects from a merMod list
}

